+++ tests/testthat/test-mable.R
@@ -69,6 +69,12 @@ test_that("mable dplyr verbs", {
     colnames %>% 
     expect_identical(c("key", "ets"))
   
+  # Test for negative tidyselect with keyed data (#120)
+  mbl_complex %>% 
+    select(-lm) %>%
+    colnames() %>% 
+    expect_identical(c("key", "ets"))
+  
   expect_output(mbl_complex %>% filter(key == "mdeaths") %>% print, "mable") %>% 
     .[["key"]] %>% 
     expect_identical("mdeaths")